Custom responces implemented for api.

+++ app/Http/Controllers/AirportController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Cache;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Validator;
 use App\Models\Airport;
 use App\Http\Resources\AirportResource;
@@ -20,7 +21,7 @@ class AirportController extends Controller
     public function index()
     {
         return AirportResource::collection(Cache::remember('airports', 60 * 60 * 24, function() {
-            return Airport::all();
+            return Airport::paginate(15);
         }));
     }
 
@@ -39,12 +40,24 @@ class AirportController extends Controller
 
         //Returning error string instead of an error message
         if ($validator->fails()) {
-            //$error = $validator->messages()->get('*');
-            return $validator->messages()->get('*');
-        } else {
-            //Validator success
-            return Airport::create($request->all());
+            $error = $validator->messages()->get('*');
+            $response = [
+                'message' => 'Validation failed.',
+                'errors' => $error,
+                'status' => 400
+            ];
+            return response($response, 400);
         }
+        //Validator success
+        $airport = new AirportResource(Airport::create($request->all()));
+
+        $response = [
+            'airport' => $airport,
+            'message' => 'Airport created successfully.',
+            'status' => '201'
+        ];
+
+        return response($response, 201);
     }
 
     /**
@@ -55,7 +68,28 @@ class AirportController extends Controller
      */
     public function show($id)
     {
-        return new AirportResource(Airport::findOrFail($id));
+        $raw_airport = Airport::find($id);
+
+        //Testing if airport exsists in database
+        if($raw_airport == null) {
+            $response = [
+                'message' => 'Airport not found.',
+                'status' => 404
+            ];
+
+            return response($response, 404);
+        }
+
+        //Passing the airport object through the airport resource for formatting
+        $airport = new AirportResource(Airport::find($id));
+        
+        $response = [
+            'airport' => $airport,
+            'message' => 'Airport found successfully',
+            'status' => 200
+        ];
+
+        return response($response, 200);
     }
 
     /**
@@ -71,13 +105,41 @@ class AirportController extends Controller
             'location' => 'required'
         ]);
 
+        //Testing data for validation
         if($validator->fails()) {
-            return $validator->messages()->get('*');
+            $error = $validator->messages()->get('*');
+            $response = [
+                'message' => 'Airport Validation failed.',
+                'errors' => $error,
+                'status' => 400
+            ];
+
+            return response($response, 400);
         } else {
+            $raw_airport = Airport::find($id);
+
+            //Testing if airport is present in database
+            if($raw_airport == null) {
+                $response = [
+                    'message' => 'Airport not found.',
+                    'status' => 404
+                ];
+
+                return response($response, 404);
+            }
+
             $airport = Airport::find($id);
             $airport->update($request->all());
             $airport->save();
-            return $airport;
+
+            $pretty_airport = new AirportResource(Airport::find($id));
+
+            $response = [
+                'airport' => $pretty_airport,
+                'message' => 'Airport updated successfully',
+                'status' => 200
+            ];
+            return response($response, 200);
         }
     }
 
@@ -89,6 +151,25 @@ class AirportController extends Controller
      */
     public function destroy($id)
     {
-        return Airport::destroy($id);
+        $raw_airport = Airport::find($id);
+
+        //Testing if airport is present in database
+        if($raw_airport == null) {
+            $response = [
+                'message' => 'Airport not found.',
+                'status' => 404
+            ];
+
+            return response($response, 404);
+        }
+
+        Airport::destroy($id);
+
+        $response = [
+            'message' => 'Airport found and deleted successfully.',
+            'status' => 200
+        ];
+
+        return response($response, 200);
     }
 }
