+++ app/Http/Controllers/FlightController.php
@@ -7,7 +7,7 @@ use Illuminate\Support\Facades\Cache;
 use App\Models\Flight;
 use App\Http\Resources\FlightResource;
 use Illuminate\Support\Facades\Validator;
-use DB;
+use Illuminate\Http\Response;
 
 class FlightController extends Controller
 {
@@ -19,7 +19,7 @@ class FlightController extends Controller
     public function index()
     {
         return FlightResource::collection(Cache::remember('flights', 60 * 60 * 12, function() {
-            return Flight::all();
+            return Flight::paginate(15);
         }));
     }
 
@@ -42,10 +42,29 @@ class FlightController extends Controller
         ]);
 
         if($validator->fails()) {
-            return $validator->messages()->get('*');
-        } else {
-            return Flight::create($request->all());
+            $error = $validator->messages()->get('*');
+            $response = [
+                'message' => 'Validation failed',
+                'errors' => $error,
+                'status' => 400
+            ];
+
+            return response($response, 400);
         }
+        
+
+        $flight = Flight::create($request->all());
+        $id = $flight->id;
+
+        $pretty_flight = new FlightResource(Flight::find($id));
+
+        $response = [
+            'flight' => $pretty_flight,
+            'message' => 'Flight created successfully.',
+            'status' => 201
+        ];
+        
+        return response($response, 201);
     }
 
     /**
@@ -56,7 +75,27 @@ class FlightController extends Controller
      */
     public function show($id)
     {
-        return new FlightResource(Flight::findOrFail($id));
+        $raw_flight = Flight::find($id);
+
+        //Testing if flight exsists in database
+        if($raw_flight == null) {
+            $response = [
+                'message' => 'Flight not found.',
+                'status' => 404
+            ];
+
+            return response($response, 404);
+        }
+
+        $flight = new FlightResource(Flight::find($id));
+
+        $response = [
+            'flight' => $flight,
+            'message' => 'Flight found.',
+            'status' => 200
+        ];
+
+        return response($response, 200);
     }
 
     /**
@@ -68,7 +107,7 @@ class FlightController extends Controller
      */
     public function update(Request $request, $id)
     {
-                //Validation
+        //Validation
         $validator = Validator::make($request->all(), [
             'departure_date',
             'arrival_date',
@@ -79,13 +118,37 @@ class FlightController extends Controller
         ]);
 
         if($validator->fails()) {
-            return $validator->messages()->get('*');
-        } else {
-            return Flight::create($request->all());
-            $flight = Flight::find($id);
-            $flight->update($request->all());
-            return $flight;
+            $errors = $validator->messages()->get('*');
+            $response = [
+                'message' => 'Validation failed',
+                'errors' => $errors,
+                'status' => 400
+            ];
+            return response($response, 400);
         }
+
+        $raw_flight = Flight::find($id);
+
+        if($raw_flight == null) {
+            $response = [
+                'message' => 'Flight not found.',
+                'status' => 404
+            ];
+
+            return response($response, 404);
+        }
+
+        $raw_flight->update($request->all());
+
+        $pretty_flight = new FlightResource(Flight::find($id));
+
+        $response = [
+            'flight' => $pretty_flight,
+            'message' => 'Flight created successfully',
+            'status' => 200
+        ];
+
+        return response($response, 200);
     }
 
     /**
@@ -96,6 +159,25 @@ class FlightController extends Controller
      */
     public function destroy($id)
     {
-        return Flight::destroy($id);
+        $raw_flight = Flight::find($id);
+
+        //Testing if flight is present in database
+        if($raw_flight == null) {
+            $response = [
+                'message' => 'Flight not found.',
+                'status' => 404
+            ];
+
+            return response($response, 404);
+        }
+
+        Flight::destroy($id);
+
+        $response = [
+            'message' => 'Flight found and deleted successfully.',
+            'status' => 200
+        ];
+
+        return response($response, 200);
     }
 }
