+++ app/Http/Controllers/PlaneController.php
@@ -7,6 +7,7 @@ use Illuminate\Support\Facades\Cache;
 use App\Models\Plane;
 use App\Http\Resources\PlaneResource;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Http\Response;
 use DB;
 
 class PlaneController extends Controller
@@ -19,7 +20,7 @@ class PlaneController extends Controller
     public function index()
     {
         return PlaneResource::collection(Cache::remember('planes', 60 * 60 * 24, function() {
-            return Plane::all();
+            return Plane::paginate(15);
         }));
     }
 
@@ -34,15 +35,35 @@ class PlaneController extends Controller
         //Validation
         $validator = Validator::make($request->all(), [
             'model' => 'required',
+            'capacity' => 'required',
             'speed' => 'required',
-            'capacity' => 'required'
+            'location' => 'required'
         ]);
 
         if($validator->fails()) {
-            return $validator->messages()->get('*');
-        } else {
-            return Plane::create($request->all());
+            $error = $validator->messages()->get('*');
+            $response = [
+                'message' => 'Validation failed',
+                'errors' => $error,
+                'status' => 400
+            ];
+
+            return response($response, 400);
         }
+        
+
+        $plane = Plane::create($request->all());
+        $id = $plane->id;
+
+        $pretty_plane = new PlaneResource(Plane::find($id));
+
+        $response = [
+            'plane' => $pretty_plane,
+            'message' => 'Plane created successfully.',
+            'status' => 201
+        ];
+        
+        return response($response, 201);
     }
 
     /**
@@ -53,7 +74,27 @@ class PlaneController extends Controller
      */
     public function show($id)
     {
-        return new PlaneResource(Plane::findOrFail($id));
+        $raw_plane = Plane::find($id);
+
+        //Testing if plane exsists in database
+        if($raw_plane == null) {
+            $response = [
+                'message' => 'Plane not found.',
+                'status' => 404
+            ];
+
+            return response($response, 404);
+        }
+
+        $plane = new PlaneResource(Plane::find($id));
+
+        $response = [
+            'plane' => $plane,
+            'message' => 'Plane found.',
+            'status' => 200
+        ];
+
+        return response($response, 200);
     }
 
     /**
@@ -69,16 +110,42 @@ class PlaneController extends Controller
         $validator = Validator::make($request->all(), [
             'model',
             'speed',
-            'capacity'
+            'capacity',
+            'airport_id',
         ]);
 
         if($validator->fails()) {
-            return $validator->messages()->get('*');
-        } else {
-            $plane = Plane::find($id);
-            $plane->update($request->all());
-            return $plane;
+            $errors = $validator->messages()->get('*');
+            $response = [
+                'message' => 'Validation failed',
+                'errors' => $errors,
+                'status' => 400
+            ];
+            return response($response, 400);
         }
+
+        $raw_plane = Plane::find($id);
+
+        if($raw_plane == null) {
+            $response = [
+                'message' => 'Plane not found.',
+                'status' => 404
+            ];
+
+            return response($response, 404);
+        }
+
+        $raw_plane->update($request->all());
+
+        $pretty_plane = new PlaneResource(Plane::find($id));
+
+        $response = [
+            'plane' => $pretty_plane,
+            'message' => 'Plane created successfully',
+            'status' => 200
+        ];
+
+        return response($response, 200);
         
     }
 
@@ -90,6 +157,25 @@ class PlaneController extends Controller
      */
     public function destroy($id)
     {
-        return Plane::destroy($id);
+        $raw_plane = Plane::find($id);
+
+        //Testing if plane is present in database
+        if($raw_plane == null) {
+            $response = [
+                'message' => 'Plane not found.',
+                'status' => 404
+            ];
+
+            return response($response, 404);
+        }
+
+        Plane::destroy($id);
+
+        $response = [
+            'message' => 'Plane found and deleted successfully.',
+            'status' => 200
+        ];
+
+        return response($response, 200);
     }
 }
