+++ app/Http/Controllers/StaffController.php
@@ -7,6 +7,7 @@ use Illuminate\Support\Facades\Cache;
 use App\Models\Staff;
 use App\Http\Resources\StaffResource;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Http\Response;
 use DB;
 
 class StaffController extends Controller
@@ -19,7 +20,7 @@ class StaffController extends Controller
     public function index()
     {
         return StaffResource::collection(Cache::remember('staff', 60 * 60 * 24, function() {
-            return Staff::all();
+            return Staff::paginate(15);
         }));
     }
 
@@ -34,16 +35,36 @@ class StaffController extends Controller
         //Validation
         $validator = Validator::make($request->all(), [
             'customer_id' => 'required',
+            'airport_id' => 'required',
             'hire_date' => 'required',
             'job' => 'required',
-            'qualifications' => 'nullable'
+            'qualifications' => 'required'
         ]);
 
         if($validator->fails()) {
-            return $validator->messages()->get('*');
-        } else {
-            return Staff::create($request->all());
+            $error = $validator->messages()->get('*');
+            $response = [
+                'message' => 'Validation failed',
+                'errors' => $error,
+                'status' => 400
+            ];
+
+            return response($response, 400);
         }
+        
+
+        $staff = Staff::create($request->all());
+        $id = $staff->id;
+
+        $pretty_staff = new StaffResource(Staff::find($id));
+
+        $response = [
+            'staff' => $pretty_staff,
+            'message' => 'Staff created successfully.',
+            'status' => 201
+        ];
+        
+        return response($response, 201);
     }
 
     /**
@@ -54,7 +75,27 @@ class StaffController extends Controller
      */
     public function show($id)
     {
-        return new StaffResource(Staff::findOrFail($id));
+        $raw_staff = Staff::find($id);
+
+        //Testing if staff member exsists in database
+        if($raw_staff == null) {
+            $response = [
+                'message' => 'Staff member not found.',
+                'status' => 404
+            ];
+
+            return response($response, 404);
+        }
+
+        $staff = new StaffResource(Staff::find($id));
+
+        $response = [
+            'staff' => $staff,
+            'message' => 'Staff memeber found.',
+            'status' => 200
+        ];
+
+        return response($response, 200);
     }
 
     /**
@@ -68,17 +109,45 @@ class StaffController extends Controller
     {
         //Validation
         $validator = Validator::make($request->all(), [
-            'hire_date' => 'required',
-            'job' => 'required',
-            'qualifications' => 'nullable'
+            'customer_id',
+            'airport_id',
+            'job',
+            'qualifications',
+            'hire_date'
         ]);
 
         if($validator->fails()) {
-            return $validator->messages()->get('*');
-        } else {
-            $staff = Staff::find($id);
-            $staff->update($request::all());
+            $errors = $validator->messages()->get('*');
+            $response = [
+                'message' => 'Validation failed',
+                'errors' => $errors,
+                'status' => 400
+            ];
+            return response($response, 400);
         }
+
+        $raw_staff = staff::find($id);
+
+        if($raw_staff == null) {
+            $response = [
+                'message' => 'Staff not found.',
+                'status' => 404
+            ];
+
+            return response($response, 404);
+        }
+
+        $raw_staff->update($request->all());
+
+        $pretty_staff = new StaffResource(Staff::find($id));
+
+        $response = [
+            'staff' => $pretty_staff,
+            'message' => 'Staff created successfully',
+            'status' => 200
+        ];
+
+        return response($response, 200);
     }
 
     /**
@@ -89,6 +158,25 @@ class StaffController extends Controller
      */
     public function destroy($id)
     {
-        return Staff::destroy($id);
+        $raw_staff = Staff::find($id);
+
+        //Testing if staff is present in database
+        if($raw_staff == null) {
+            $response = [
+                'message' => 'Staff not found.',
+                'status' => 404
+            ];
+
+            return response($response, 404);
+        }
+
+        Staff::destroy($id);
+
+        $response = [
+            'message' => 'Staff found and deleted successfully.',
+            'status' => 200
+        ];
+
+        return response($response, 200);
     }
 }
